Ninjamaster: Controllerified.
