--- deploy/www/ninjamaster/tools.php
@@ -1,7 +1,0 @@
-<?php
-
-$page = 'tools';
-$pages = array('tools'=>array('title'=>'Web Administration Tools'));
-
-
-display_static_page($page, $pages, $vars=array(), $options=array('private'=>false));
